Implemented tsdoc, updated RMB class listen and cast method
