+++ src/lib/redditmarketbot.ts
@@ -1,32 +1,76 @@
-import { sleep, fetchPosts, castNotify } from './actions';
+import { sleep, fetchPosts } from './actions';
 import { PostI, PostChildrenI, ParamsI, RedditMarketBotI } from '../types/types';
+import * as googleTTS from 'google-tts-api';
+import ChromecastAPI from 'chromecast-api';
+import Device from 'chromecast-api/lib/device';
 
 class RedditMarketBot implements RedditMarketBotI {
-  params;
+  params: ParamsI;
+  matches: number;
 
-  constructor(params: ParamsI) {
+  constructor(params: ParamsI, matches: number = 0) {
     this.params = params;
+    this.matches = matches;
   }
 
-  async listen() {
+  /**
+   * Scans subreddit and alerts user via console/casting on new posts that match the search query
+   *
+   * @remarks
+   * Need to implement pagination
+   *
+   * @beta
+   */
+  async listen(callback: () => void) {
     while (true) {
-      let matches = 0;
-      const data: PostI = await fetchPosts(this.params.subreddit, this.params.postCount);
+      this.matches = 0;
+      const data: PostI = await fetchPosts(this.params.subreddit, this.params.postLimit);
 
       data.data.children.forEach((post: PostChildrenI) => {
         if (post.data.title.indexOf(this.params.query) !== -1) {
           console.log('\x1b[36m%s\x1b[0m', post.data.title);
           console.log('\x1b[32m%s\x1b[0m', post.data.url + '\n');
-          matches++;
+          this.matches++;
         }
       });
 
-      if (this.params.enableCasting && matches > 0) {
-        await castNotify(matches, this.params.subreddit);
-      }
-
+      callback();
       await sleep(30000);
     }
+  }
+
+  /**
+   * Gets the number of match (if any), converts it into audio, then casts it into a Chromecast device
+   *
+   * @remarks
+   * Need to implement manual device selection
+   *
+   * @beta
+   */
+  cast() {
+    const msg: string =
+      this.matches.toString() +
+      `${this.matches === 1 ? ' match' : ' matches'}` +
+      ' have been found on ' +
+      this.params.subreddit;
+
+    const url: string = googleTTS.getAudioUrl(msg, {
+      lang: 'en',
+      slow: false,
+      host: 'https://translate.google.com'
+    });
+
+    const client = new ChromecastAPI();
+
+    client.on('device', (device: Device) => {
+      device.play(url, (err) => {
+        if (!err) console.log('\nPlaying notification on ' + device.friendlyName + '.');
+      });
+
+      device.on('finished', () => {
+        device.close();
+      });
+    });
   }
 }
